requirements.txt
flask==3.0.0
requests==2.31.0
torch==2.1.0
numpy==1.26.1
pillow==10.1.0
python-dotenv==1.0.0
